Add finished non graphics version of the game
Created user selection function where the player is prompted
to enter "Rock" "Paper" or "Scissors".
Added error message if the player spelled their answer wrong.

Added message notifying player they canceled the game.

This function is case insensitive and always capitalizes the first
letter and makes the rest of the players word lowercase no matter
how it was spelled which makes it easier for the game to know what
the player entered.

Added playRound function which compares the choices of both the
player and the computer and announces the results in an alert box.

Added a playAgain function which will ask the player to play again
once the results from the previous round are in.