Edit the game to work with one main function call
Moved getPlayerChoice and getComputerChoice functions to be called
at the top of the playRound function

Made it so points get added to playerScore and computerScore
variables when either the player or the computer wins a round

Edited the playAgain function to only reload the page on confirm
or say "Ok, bye!" if canceled.

Created a game() function with a for loop that will play 5 rounds
It will first prompt you to enter a name
This will store your name to be used if the player wins
This function keeps track of the player and computers score
After 5 rounds the game ends and an alert pops up saying who won
and what the score was.
If there's a tie, an alert will pop up and say that it was a tie
game with a score of x-x.

There is an issue where if the player hits cancel the game will
continue playing until 5 actions are completed

I would like to figure out how to fix this issue in the future.